Simplified Examples Pages  - upgrade caddy  - Seperated directories used for bower  - separate tests from examples
