+++ examples/column-chart-examples.html
@@ -1,0 +1,44 @@
+<html>
+<head>
+    <meta charset="utf-8"/>
+    <link rel="stylesheet" href="styles/main.css"/>
+    <link rel="stylesheet" href="styles/demo.css"/>
+    <title>Columns</title>
+</head>
+<body class="ft">
+<style>
+    svg {
+        margin:5px;
+    }
+</style>
+<script src=scripts/vendor.js></script>
+<script src=scripts/o-charts.js></script>
+
+<div id="views"></div>
+<script>
+    var oCharts = require('o-charts');
+    var d3 = require('d3');
+
+    d3.select('#views').append('div').data([{
+        comment: 'Column chart',
+        title: 'Columns: ',
+        subtitle: 'Categories, with a horizontal independent axis',
+        dependentAxisOrient: 'left',
+        independentAxisOrient: 'bottom',
+        hideSource: true,
+        x: { series: 'key' },
+        y: { series: ['value', 'value2'] },
+        data: [
+            { key: 'red', value:      0.583, value2:      1.583},
+            { key: 'blue', value: 0.12, value2:  2},
+            { key: 'green machine', value: 1.03, value2:  1.4},
+            { key: 'purple', value:     1.348, value2:  1.9}
+        ],
+        stack:true,
+        dataType : 'categorical'
+    }]).call(oCharts.chart.column);
+</script>
+
+
+</body>
+</html>
