+++ examples/index.html
@@ -1,0 +1,28 @@
+<!DOCTYPE html>
+<html>
+<head>
+    <meta charset="utf-8"/>
+    <link rel="stylesheet" href="styles/main.css"/>
+    <link rel="stylesheet" href="styles/demo.css"/>
+    <title>o-chart Examples</title>
+</head>
+<body class="ft">
+
+<h1>o-chart Examples</h1>
+
+<h2>Axes</h2>
+<ul>
+    <li><a href="category-axes-examples.html">Category axes</a></li>
+    <li><a href="date-axes-examples.html">Date axes</a></li>
+    <li><a href="number-axes-examples.html">Number axes</a></li>
+</ul>
+
+<h2>Charts</h2>
+<ul>
+    <li><a href="column-chart-examples.html">Column chart</a></li>
+    <li><a href="bar-chart-examples.html">Bar chart</a></li>
+    <li><a href="line-chart-examples.html">Line chart</a></li>
+</ul>
+
+</body>
+</html>
