+++ examples/number-axes-examples.html
@@ -1,0 +1,126 @@
+<!DOCTYPE html>
+<html>
+<head>
+    <meta charset="utf-8"/>
+    <link rel="stylesheet" href="styles/main.css"/>
+    <link rel="stylesheet" href="styles/demo.css"/>
+    <title>Number Axes</title>
+</head>
+<body class="ft">
+<style>
+    div { clear: both; }
+</style>
+<h1>Number axes</h1>
+
+
+<script src="scripts/vendor.js"></script>
+<script src="scripts/o-charts.js"></script>
+
+<div id=views></div>
+<script>
+    var oCharts = require('o-charts');
+    var d3 = require('d3');
+
+    var margin = {
+        top: 20, left: 50, bottom: 70, right: 50
+    }
+    var data = {
+        orient: "left",
+        simple: false,
+        scale: d3.scale.linear()
+            .range([0, 400])
+            .domain([11.2, 7])
+    };
+
+    var divs = d3.select('#views');
+
+    divs.append('h2')
+        .text("Number Axis");
+
+    divs.append('svg')
+        .data([data])
+        .attr('width', function (d) {
+            if (d.orient == "bottom") {
+                var r = d.scale.range();
+                return r[0] + r[1] + margin.bottom
+            }
+            return margin.left + margin.right
+        })
+        .attr('class', 'ft-chart')
+        .attr('height', function (d) {
+            if (d.orient == "bottom") {
+                return margin.bottom
+            }
+            var r = d.scale.range();
+            return r[0] + r[1] + margin.bottom
+        })
+        .each(function (d, i) {
+            var axis = oCharts.axis.number()
+                .simple(d.simple)
+                .scale(d.scale)
+                .orient(d.orient);
+
+            d3.select(this)
+                .append('g')
+                .attr('transform', 'translate(' + margin.left + ',' + margin.top + ')')
+                .call(axis);
+        });
+
+
+</script>
+
+
+<div id=horizontal></div>
+<script>
+    var oCharts = require('o-charts');
+    var d3 = require('d3');
+
+    var margin = {
+        top: 20, left: 50, bottom: 70, right: 50
+    }
+    var data = {
+        orient: "bottom",
+        simple: false,
+        scale: d3.scale.linear()
+            .range([0, 400])
+            .domain([11.2, 7])
+    };
+
+    var divs = d3.select('#horizontal');
+
+    divs.append('h2')
+        .text("Number Axis");
+
+    divs.append('svg')
+        .data([data])
+        .attr('width', function (d) {
+            if (d.orient == "bottom") {
+                var r = d.scale.range();
+                return r[0] + r[1] + margin.bottom
+            }
+            return margin.left + margin.right
+        })
+        .attr('class', 'ft-chart')
+        .attr('height', function (d) {
+            if (d.orient == "bottom") {
+                return margin.bottom
+            }
+            var r = d.scale.range();
+            return r[0] + r[1] + margin.bottom
+        })
+        .each(function (d, i) {
+            var axis = oCharts.axis.number()
+                .simple(d.simple)
+                .scale(d.scale)
+                .orient(d.orient);
+
+            d3.select(this)
+                .append('g')
+                .attr('transform', 'translate(' + margin.left + ',' + margin.top + ')')
+                .call(axis);
+        });
+
+
+</script>
+</body>
+</html>
