+++ test/fixtures/bar-chart.html
