+++ test/fixtures/blank-chart.html
