+++ test/fixtures/scripts/category-axes.js
@@ -1,5 +1,4 @@
-
-var oCharts = require('../../src/scripts/o-charts');
+var oCharts = require('../../../src/scripts/o-charts');
 var dateUtils = oCharts.util.dates;
 var d3 = require('d3');
