+++ test/fixtures/scripts/line-chart.js
@@ -1,4 +1,4 @@
-var oCharts = require('../../src/scripts/o-charts');
+var oCharts = require('../../../src/scripts/o-charts');
 var d3 = require('d3');
 
 var y = [   { series: ['value', 'value2', 'value3'] },
