+++ test/fixtures/scripts/series-key.js
@@ -1,5 +1,5 @@
 
-var oCharts = require('../../src/scripts/o-charts');
+var oCharts = require('../../../src/scripts/o-charts');
 var d3 = require('d3');
 
 var lineData = [
