+++ test/helper.js
@@ -11,11 +11,12 @@ function appendScript(fileObj){
     document.body.appendChild(link)
 }
 function loadAssets(page){
-    document.body.innerHTML = __html__['_site/' + (page || 'index') + '.html'];
-    appendCSS({path: '_site/styles/demo.css'});
-    appendCSS({path: '_site/styles/main.css'});
+    document.body.innerHTML = __html__['_test/' + (page || 'index') + '.html'];
+    //appendCSS({path: '_test/styles/demo.css'});
+    //appendCSS({path: '_test/styles/main.css'});
     appendScript({path: '_site/scripts/vendor.js'});
-    appendScript({path: '_site/scripts/' + (page || 'o-charts') + '.js'});
+    !page && appendScript({path: '_site/scripts/o-charts.js'});
+    page && appendScript({path: '_test/scripts/' + (page ) + '.js'});
 }
 
 module.exports = {
